Extract shared chat-service transport and helpers
