+++ Modules/Sources/WooAIAssistant/Networking/JetpackAIQueryClient+Localization.swift
@@ -2,11 +2,6 @@ import Foundation
 
 extension JetpackAIQueryClient {
     enum Localization {
-        static let nonHTTPResponse = NSLocalizedString(
-            "ai.assistant.networking.non_http_response",
-            value: "The assistant returned an unexpected response.",
-            comment: "Surfaced when the chat transport receives a non-HTTP URLResponse from the WPCOM proxy."
-        )
         static let httpFailureFallback = NSLocalizedString(
             "ai.assistant.networking.http_failure",
             value: "The assistant returned an error (HTTP %1$d).",
